Reorganised the template outputs for component. Moved Connected component to a Plugin definition file which will act as root component for the Plugin.
