+++ template/src/components/toggleButton/index.js
@@ -1,0 +1,4 @@
+import ToggleButton from './ToggleButton';
+
+export default ToggleButton;
+export { ToggleButton };
