+++ template/src/components/{{name}}/index.js
@@ -1,0 +1,4 @@
+import {{pascalcase name}} from './{{name}}';
+
+export default {{pascalcase name}};
+export { {{pascalcase name}} };
